Publish pattern (!p) has been removed from the language completely. Its use would freqently give rise to subtle and frustrating bugs.
Type assertions are now available; write :!: T to assert T as the type of an expression or the return type of a definition.  Unlike a type ascription :: T, the typechecker will assume that the assertion :!: T is correct, and will perform no typechecking on the underyling expression. Note that this _will_ compromise the typechecker if used incorrectly; it should be used only for prototyping purposes.