Experimental support for isolated regions (protected from forced termination). In more detail: - parser and AST include "isolated" keyword - nodes Isolate and Unisolate are used to put tokens into an isolated region and remove them - GroupCell was generalized to Group, which includes just those aspects of GroupCell related to forcibly terminating tokens. An isolated region uses a special IsolatedGroup which ignores forcible termination. - IsolatedRegion was added to attach the isolated region to the root execution so it won't be garbage collected even if it is cut off higher up.
It's not clear to me that the current region/group distinction is appropriate or correct. When a group is forcibly terminated, when and how should this affect transactions and otherwise combinators inside the group? If some tokens are still pending at sites, does this affect the answer? I need to discuss with David.