Type inference will now guess a type parameter when there is no minimal type, and report a warning when it makes such a guess. Ref(0) will now pass the typechecker without annotations (it correctly guesses Integer). This fixes issue #103.
This commit also fixes a hidden bug in the typechecker regarding function arity checking in the context of type parameter inference.

util.inc has been modified to make invokeExecutable type safe, but it still does not pass the typechecker due to a strange control flow bug. I am opening a separate issue for that problem.