Make spawns only occur when the stack is "full".
This turns out to be a really important optimization. Gaining a factor
of 6 in one case (Filling an array with numbers). 

This technique is a combination of TCO and sequentialization. We use the
scheduler as the trampoline. This allows calls back from sites to Orc
code to occur on the stack without the scheduler.