Fix a race in the interpreter that was unmasked by the new scheduler.
This commit fixes the following race condition:

A token is in a group which is killed. The group schedules the token for clean up. The token simultaneously moves to a subgroup. The token is not double scheduled and in a group that has not been killed (it has been scheduled to be killed just like the token, but isKilled does not yet return true). So the token can execute normally except that it is in the queues twice so it can run concurrently with itself. This can cause a later crash (most often a "halted token scheduled" error). However, the token can only execute until the subgroup executes and becomes killed. In the past this was never very long, but with the new work-stealing scheduler can result in the subgroup waiting for the token to complete before the subgroup can run (if the token is on top of the work stack for a thread). So this happens much more often.