Eliminate some really deep recursion caused by reduceRight.
This replaces it with mkString which is iterative. The problem
was that the GraalVM seems to have a lot of trouble with deep
recursion or something else about how the list of was being
processed. In addition, the old version was concatinating 1000s
of strings instead of using a string builder.