PorcE: Occationally reset ObjectOutputStreams. Weaken map.
This is a work around for the deeper problem that OOS keeps a
strong reference to every object that it has ever serialized.
The memory isn't actually the problem, instead the number of
objects in the internal table makes look-ups on that table slow
and this slows down the whole program significantly.