Flag futures bound to expressions that publish zero or one times.
When constructing Porc Future AST nodes, check Orctimizer's
PublicationCountAnalysis for the expression that the future is bound to.
If the expression publishes once or fewer times, flag the Porc future as
a race-free resolution.  (I.e., when a publication arrives to be bound
to the flagged future, the proposed resolution value has no competing
publications -- it will always be the only value to which that future is
resolved.)

The runtimes will use this flag for faster resolution in this case.
(Not implemented in this commit.)