PorcE: Disable checkinit in PorcE.
This was intentionally off before because the initchecks have a
significant overhead (an atomic operation on initialization) on
every Scala field. This is noticable in PorcE in part because it
makes it a harder for Graal to eliminate the Scala objects after
Truffle partial evaluation. Checkinit also makes the Graal IR much
harder to read when I have to do that.