[PorcE DOES NOT BUILD] Update sites for new API.
This makes minimal changes to each site implementation to update
it. In many cases this is just changing imports to import the
compatibility layer. In some cases renaming some references is
also done.

Just a few sites are actually updated to use the new API.

This commit can run tiny Orc programs on the token interpreter.
It is not tested beyond "1 + 2".