PorcE: Change large stack handling.
When the stack is large unroll the whole thing instead of just
spawning the current task. This prevents running exactly the large
depth and spawning more often. Also spawn at a lower threshold
if spawning has already happened at that point. This may reduce
the number of spawns which are included in compiled code.

Overall this dramatically reduces the number of functions Graal
compiles and also improves performance (in at least some cases).